Add test for creating mastery string, and fixes that came up with it

+++ src/main/java/no/stelar7/api/l4j8/pojo/staticdata/rune/StaticRuneList.java
@@ -7,7 +7,7 @@ import java.util.*;
 public class StaticRuneList
 {
     private BasicData             basic;
-    private Map<Long, StaticRune> data;
+    private Map<Integer, StaticRune> data;
     
     
     @Override
@@ -65,7 +65,7 @@ public class StaticRuneList
      *
      * @return the data
      */
-    public Map<Long, StaticRune> getData()
+    public Map<Integer, StaticRune> getData()
     {
         return this.data;
     }
